Split settings into account and item registry
