+++ components/settings/account-settings.tsx
@@ -10,15 +10,14 @@ import { Button } from '@/components/ui/button';
 import { Badge } from '@/components/ui/badge';
 import { ScrollArea } from '@/components/ui/scroll-area';
 import type { Database } from '@/supabase/types';
-import { createInviteCodeAction, registerItemAction } from '@/app/settings/actions';
+import { createInviteCodeAction } from '@/app/account-settings/actions';
 
 interface AccountSettingsProps {
   user: User | null;
   invites: Database['public']['Tables']['invite_codes']['Row'][];
-  items: Database['public']['Tables']['items']['Row'][];
 }
 
-export function AccountSettings({ user, invites, items }: AccountSettingsProps) {
+export function AccountSettings({ user, invites }: AccountSettingsProps) {
   const router = useRouter();
   const [displayName, setDisplayName] = useState(user?.user_metadata?.full_name ?? '');
   const [inviteRole, setInviteRole] = useState('admin');
@@ -26,10 +25,6 @@ export function AccountSettings({ user, invites, items }: AccountSettingsProps) 
   const [inviteFeedback, setInviteFeedback] = useState<string | null>(null);
   const [inviteError, setInviteError] = useState<string | null>(null);
   const [creatingInvite, startCreateInvite] = useTransition();
-  const [itemId, setItemId] = useState('');
-  const [itemError, setItemError] = useState<string | null>(null);
-  const [itemFeedback, setItemFeedback] = useState<string | null>(null);
-  const [registeringItem, startRegisterItem] = useTransition();
 
   const handleInviteSubmit = (event: FormEvent<HTMLFormElement>) => {
     event.preventDefault();
@@ -54,30 +49,11 @@ export function AccountSettings({ user, invites, items }: AccountSettingsProps) 
     });
   };
 
-  const handleRegisterItem = (event: FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-    setItemError(null);
-    setItemFeedback(null);
-    const formData = new FormData();
-    formData.append('id', itemId.trim());
-
-    startRegisterItem(async () => {
-      const response = await registerItemAction(formData);
-      if (response && !response.ok) {
-        setItemError(response.error ?? 'Unable to register item');
-        return;
-      }
-      setItemFeedback(`Item ${itemId.trim()} registered.`);
-      setItemId('');
-      router.refresh();
-    });
-  };
-
   return (
     <div className="space-y-8">
       <div>
         <h1 className="text-3xl font-semibold text-white">Settings</h1>
-        <p className="text-sm text-foreground/70">Manage your account, invite new admins, and register loot items.</p>
+        <p className="text-sm text-foreground/70">Manage your account information and invite new administrators.</p>
       </div>
       <Card>
         <CardHeader className="flex flex-col gap-2 sm:flex-row sm:items-center sm:justify-between">
@@ -176,65 +152,6 @@ export function AccountSettings({ user, invites, items }: AccountSettingsProps) 
                     </td>
                     <td className="px-4 py-2 text-xs text-foreground/60">
                       {new Date(invite.created_at).toLocaleString()}
-                    </td>
-                  </tr>
-                ))}
-              </tbody>
-            </table>
-          </ScrollArea>
-        </CardContent>
-      </Card>
-
-      <Card>
-        <CardHeader className="flex flex-col gap-2 sm:flex-row sm:items-center sm:justify-between">
-          <div>
-            <CardTitle>Registered items</CardTitle>
-            <CardDescription>Items listed here can be referenced by loot tables and simulations.</CardDescription>
-          </div>
-          <Badge variant="default">{items.length} items</Badge>
-        </CardHeader>
-        <CardContent className="grid gap-6 lg:grid-cols-[320px_1fr]">
-          <form className="space-y-4 rounded-2xl border border-white/10 bg-black/40 p-4" onSubmit={handleRegisterItem}>
-            <div className="space-y-2">
-              <Label htmlFor="item-id">Item ID</Label>
-              <Input
-                id="item-id"
-                value={itemId}
-                onChange={(event) => setItemId(event.target.value)}
-                placeholder="minecraft:diamond"
-                required
-              />
-              <p className="text-xs text-foreground/50">
-                Only the canonical item identifier is stored. Display metadata is resolved in-game.
-              </p>
-            </div>
-            <Button type="submit" disabled={registeringItem}>
-              {registeringItem ? 'Registering…' : 'Register item'}
-            </Button>
-            {itemError && <p className="text-sm text-destructive">{itemError}</p>}
-            {itemFeedback && <p className="text-sm text-primary">{itemFeedback}</p>}
-          </form>
-          <ScrollArea className="h-64 rounded-2xl border border-white/10 bg-black/30">
-            <table className="w-full text-sm text-foreground/80">
-              <thead className="sticky top-0 bg-black/60 text-xs uppercase tracking-wide text-foreground/60">
-                <tr>
-                  <th className="px-4 py-2 text-left">ID</th>
-                  <th className="px-4 py-2 text-left">Registered</th>
-                </tr>
-              </thead>
-              <tbody>
-                {items.length === 0 && (
-                  <tr>
-                    <td className="px-4 py-4 text-center text-foreground/50" colSpan={2}>
-                      No items registered yet. Add at least one to start building loot tables.
-                    </td>
-                  </tr>
-                )}
-                {items.map((item) => (
-                  <tr key={item.id} className="border-b border-white/5">
-                    <td className="px-4 py-2 font-mono text-xs">{item.id}</td>
-                    <td className="px-4 py-2 text-xs text-foreground/60">
-                      Added {new Date(item.created_at).toLocaleString()}
                     </td>
                   </tr>
                 ))}
